refactor: combine receipt and receipt book util functions
